requirements.txt
nipy
numpy==1.17
scipy
nibabel
flywheel-sdk==11.3.1
matplotlib